requirements.txt
python-telegram-bot==21.3
apscheduler==3.10.4
openai==1.35.0
python-dotenv==1.0.1
requests==2.32.3
cryptography==42.0.8
fastapi==0.111.0
uvicorn==0.30.1
jinja2==3.1.4
python-multipart==0.0.9
